Removed redundant auto recovery logic in order to fix buggy idle channels spawning.
