+++ amqp/src/main/resources/amqp.conf
@@ -28,7 +28,7 @@ config.amqp = {
   # Interval of the heartbeats that are used to detect broken connections. Zero for none. See for more details: https://www.rabbitmq.com/heartbeats.html
   heartbeatIntervalSec = 30
   # Interval of connection recovery attempts. See for more details: https://www.rabbitmq.com/api-guide.html#connection-recovery
-  networkRecoveryIntervalMs = 5000
+  networkRecoveryIntervalMs = 500
   
   # Specify the size of the limit of unacknowledged messages on a queue basis
   prefetchCount = 10
